+++ packages/code-studio/src/main/AppRouter.tsx
@@ -7,12 +7,16 @@ import {
 } from 'react-router-dom';
 import AppInit from './AppInit';
 
+const basename = document.location.pathname;
 function AppRouter(): ReactElement {
   return (
-    <Router basename={import.meta.env.BASE_URL}>
+    <Router basename={basename}>
       <Switch>
         <Route exact path="/" component={AppInit} />
-        <Route path="/notebook/:notebookPath+" component={AppInit} />
+        <Route
+          path={`/${import.meta.env.VITE_ROUTE_NOTEBOOKS}:notebookPath+`}
+          component={AppInit}
+        />
         <Redirect from="*" to="/" />
       </Switch>
     </Router>