+++ packages/jsapi-components/src/index.ts
@@ -1,4 +1,8 @@
 export { default as TableInput } from './TableInput';
+export * from './RefreshTokenBootstrap';
+export * from './RefreshTokenUtils';
+export { default as useBroadcastChannel } from './useBroadcastChannel';
+export { default as useBroadcastLoginListener } from './useBroadcastLoginListener';
 export { default as useInitializeViewportData } from './useInitializeViewportData';
 export { default as useTable } from './useTable';
 export { default as useTableColumn } from './useTableColumn';