+++ packages/react-hooks/src/index.ts
@@ -1,3 +1,4 @@
+export { default as useContextOrThrow } from './useContextOrThrow';
 export { default as usePrevious } from './usePrevious';
 export { default as useForwardedRef } from './useForwardedRef';
 export { default as useCopyToClipboard } from './useCopyToClipboard';
@@ -6,3 +7,4 @@ export type {
   UsePromiseFactoryOptions,
   UsePromiseFactoryResult,
 } from './usePromiseFactory';
+export * from './useDebouncedValue';